Fixed problem with read transactions freezing the AXI bus. The problem is described in https://www.fpgarelated.com/showthread/comp.arch.fpga/127408-1.php https://groups.google.com/g/comp.arch.fpga/c/6yjytarRZDk/m/Xr4mJ9ceAgAJ
It seems that the implementation of the AXI4-Lite protocol
in the Altera Qsys Interconect is buggy.
Even though it keeps RREADY asserted in the cycle in which
it receives ARREADY, it is not able to accept RVALID in that cycle.
The RREADY may be asserted at fastest in the next cycle.